AI/requirements.txt
python-dotenv
pandas==2.3.2
numpy==2.3.2
certifi==2025.8.3
scikit-learn==1.7.1
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.20
pydantic==2.7.1

# -e .